# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Readbase",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/me/integrations/gitlab/start": {
      "get": {
        "tags": [
          "gitlab"
        ],
        "summary": "Start Gitlab Connection",
        "operationId": "start_gitlab_connection_api_me_integrations_gitlab_start_get",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/integrations/gitlab/callback": {
      "get": {
        "tags": [
          "gitlab"
        ],
        "summary": "Complete Gitlab Connection",
        "operationId": "complete_gitlab_connection_api_me_integrations_gitlab_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "readbase_gitlab_oauth_state",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Gitlab Oauth State"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/integrations/gitlab": {
      "get": {
        "tags": [
          "gitlab"
        ],
        "summary": "Gitlab Connection",
        "operationId": "gitlab_connection_api_me_integrations_gitlab_get",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitlabConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gitlab"
        ],
        "summary": "Disconnect Gitlab Connection",
        "operationId": "disconnect_gitlab_connection_api_me_integrations_gitlab_delete",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitlabConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/integrations/gitlab/projects": {
      "get": {
        "tags": [
          "gitlab"
        ],
        "summary": "Gitlab Projects",
        "operationId": "gitlab_projects_api_me_integrations_gitlab_projects_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitlabProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/gitlab/projects": {
      "get": {
        "tags": [
          "gitlab"
        ],
        "summary": "Workspace Gitlab Projects",
        "operationId": "workspace_gitlab_projects_api_workspaces__workspace_id__gitlab_projects_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitlabProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invites": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "List Invites Endpoint",
        "operationId": "list_invites_endpoint_api_invites_get",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/join/{join_token}": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "Link Invite Preview Endpoint",
        "operationId": "link_invite_preview_endpoint_api_invites_join__join_token__get",
        "parameters": [
          {
            "name": "join_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Join Token"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{invite_id}/accept": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Accept Invite Endpoint",
        "operationId": "accept_invite_endpoint_api_invites__invite_id__accept_post",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invite Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{invite_id}/reject": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Reject Invite Endpoint",
        "operationId": "reject_invite_endpoint_api_invites__invite_id__reject_post",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invite Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{invite_id}/revert": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Revert Invite Endpoint",
        "operationId": "revert_invite_endpoint_api_invites__invite_id__revert_post",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invite Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GitlabConnectionResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "gitlab_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gitlab User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "GitlabConnectionResponse"
      },
      "GitlabProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path_with_namespace": {
            "type": "string",
            "title": "Path With Namespace"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "clone_url": {
            "type": "string",
            "title": "Clone Url"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path_with_namespace",
          "web_url",
          "clone_url",
          "visibility"
        ],
        "title": "GitlabProjectResponse"
      },
      "GitlabProjectsResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/GitlabProjectResponse"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "projects"
        ],
        "title": "GitlabProjectsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceInviteListItemResponse": {
        "properties": {
          "invite_id": {
            "type": "string",
            "title": "Invite Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_name": {
            "type": "string",
            "title": "Workspace Name"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "invitee_email": {
            "type": "string",
            "title": "Invitee Email"
          },
          "invitee_name": {
            "type": "string",
            "title": "Invitee Name"
          },
          "invitee_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invitee User Id"
          },
          "invitor_user_id": {
            "type": "string",
            "title": "Invitor User Id"
          },
          "invitor_name": {
            "type": "string",
            "title": "Invitor Name"
          },
          "invitor_designation": {
            "type": "string",
            "title": "Invitor Designation"
          },
          "relation": {
            "type": "string",
            "title": "Relation"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "node_display_name": {
            "type": "string",
            "title": "Node Display Name"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "can_accept": {
            "type": "boolean",
            "title": "Can Accept",
            "default": false
          },
          "can_reject": {
            "type": "boolean",
            "title": "Can Reject",
            "default": false
          },
          "can_revert": {
            "type": "boolean",
            "title": "Can Revert",
            "default": false
          },
          "invite_method": {
            "type": "string",
            "title": "Invite Method",
            "default": "email"
          },
          "join_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Join Token"
          },
          "join_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Join Path"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "invite_id",
          "workspace_id",
          "workspace_name",
          "direction",
          "invitee_email",
          "invitee_name",
          "invitor_user_id",
          "invitor_name",
          "invitor_designation",
          "relation",
          "reason",
          "node_display_name",
          "status",
          "created_at"
        ],
        "title": "WorkspaceInviteListItemResponse"
      },
...
      "WorkspaceInvitesResponse": {
        "properties": {
          "received": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
            },
            "type": "array",
            "title": "Received"
          },
          "sent": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceInviteListItemResponse"
            },
            "type": "array",
            "title": "Sent"
          }
        },
        "type": "object",
        "required": [
          "received",
          "sent"
        ],
        "title": "WorkspaceInvitesResponse"
      },
...
    }
  }
}